+++ src/dto/accounting/PagedReportByDateDto.cs
@@ -18,13 +18,13 @@ public class PagedReportByDateDto : PagedResult<TransactionReportDto>
         summary.endDate = end.toUTC6();
     }
 
-    public void setValidTransactionData((int quantity, double total) value)
+    public void setValidTransactionData((int quantity, decimal total) value)
     {
         summary.validQuantity = value.quantity;
         summary.validTotal = value.total;
     }
 
-    public void setInvalidTransactionData((int quantity, double total) value)
+    public void setInvalidTransactionData((int quantity, decimal total) value)
     {
         summary.invalidQuantity = value.quantity;
         summary.invalidTotal = value.total;
